refactor(admin): extract AdminInfraService — drop infra controller's db import
Item C 2/5. AdminInfraController imported SurrealService from src/db
(import/no-restricted-paths disable + TODO). Its two DB touchpoints — the
connection ping in healthComponents() and the per-tenant migration audit
in migrations() — move into a new AdminInfraService (pingDb +
migrationsAudit). The controller injects it (constructor 8→7 params, also
shedding the now-unused ApiKeyService) and delegates; src/db import +
disable removed.

Behaviour-identical. The four admin contract unit specs that constructed
the controller directly are updated to build it through AdminInfraService
(the typecheck gate from #44 caught all four signature breaks — working as
intended). typecheck + lint clean; 700 unit + 128 e2e green.
