+++ test/contracts-admin-now.unit-spec.ts
@@ -29,7 +29,6 @@ function makeController(): AdminInfraController {
     undef,
     undef,
     undef,
-    undef,
     activity,
     undef,
     undef,
